+++ src/DataType/InvoiceData.php
@@ -47,4 +47,16 @@ class InvoiceData implements InvoiceDataInterface
     {
         return $this->invoiceConfiguration;
     }
+
+    public function getInvoiceFilename(): string
+    {
+        $filename = 'invoice';
+
+        $invoiceNumber = $this->invoiceConfiguration->getNumber();
+        if ($invoiceNumber) {
+            $filename .= '-' . $invoiceNumber;
+        }
+
+        return $filename . '.pdf';
+    }
 }